requirements-dev.txt
-r requirements.txt
coverage==5.5
coveralls==3.2.0
cssselect==1.1.0
django-debug-toolbar==3.2.2
django-nose==1.4.7
django-webtest==1.9.7
docopt==0.6.2
ecdsa==0.17.0
Fabric==2.6.0
factory-boy==3.2.0
lxml==4.6.3
nose==1.3.7
paramiko==2.7.2
path.py==12.5.0
pycrypto==2.6.1
pyquery==1.4.3
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
waitress==2.0.0
WebOb==1.8.7
WebTest==3.0.0
